Refactor : Add PayMachine class
- Refactor : Add PayMachine class to seperate main()
